+++ src/components/HomepageFeatures/Feature/Feature.module.css
@@ -1,0 +1,4 @@
+.svg {
+    height: 200px !important;
+    width: 30% !important;
+}
