+++ src/components/HomepageFeatures/Feature/index.ts
@@ -1,0 +1,3 @@
+import Feature from './Feature';
+
+export default Feature;
