+++ src/components/HomepageFeatures/FeatureSource/index.ts
@@ -1,0 +1,3 @@
+import FeaturesSource from "./FeatureSource";
+
+export default FeaturesSource;
